+++ Onur.Tests/Using.cs
@@ -13,4 +13,4 @@
 * along with onur. If not, see <https://www.gnu.org/licenses/>.
 */
 
-global using System.IO;
+global using Xunit;